fix(avatar) : AvatarDTO가 텍스처를 포함하도록 수정 #8
